pyproject.toml
[project]
name = "llm-controlled-car"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyserial",
    "bleak",
    "wavio",
    "sounddevice",
    "faster-whisper",
    "TTS",
    "ollama"
]
